Adding confirmation to the deleting of products and categories

+++ delete_category.php
@@ -1,18 +1,34 @@
 <?php
-	$title = "Deleting category";
 	require("header_footer/header.php");
-	$squery = 'DELETE FROM categories WHERE `categories`.`id`='.$_GET['del_id'];
-	$fquery = 'DELETE FROM categories_products WHERE `categories_products`.`category_id`='.$_GET['del_id'];
-	if(!mysqli_query($con,$fquery))
+	if(!isset($_GET['confirmation']))
 	{
-		echo 'Error: '.mysqli_error($con);
+		echo '<center><h1>Do you really want to delete this category?</h1></center>';
+		?>
+		<center><h3><a href=<?php echo $delete_category_path.'?del_id='.$_GET['del_id'].'&confirmation=1' ?> >Yes</a></h3></center>
+		<center><h3><a href=<?php echo $delete_category_path.'?del_id='.$_GET['del_id'].'&confirmation=0' ?> >No</a></h3></center>
+		<?php
 	}
-	if(!mysqli_query($con,$squery))
+	else if($_GET['confirmation'])
 	{
-		echo 'Error: '.mysqli_error($con);
-	}
+		$title = "Deleting category";
+		require("header_footer/header.php");
+		$squery = 'DELETE FROM categories WHERE `categories`.`id`='.$_GET['del_id'];
+		$fquery = 'DELETE FROM categories_products WHERE `categories_products`.`category_id`='.$_GET['del_id'];
+		if(!mysqli_query($con,$fquery))
+		{
+			echo 'Error: '.mysqli_error($con);
+		}
+		if(!mysqli_query($con,$squery))
+		{
+			echo 'Error: '.mysqli_error($con);
+		}
 
-	header("Location: ".$list_categories_path);
+		header("Location: ".$list_categories_path);
+	}
+	else
+	{
+		header("Location: ".$list_categories_path);
+	}
 
 ?>
